# api/v1/api.openapi.yaml
openapi: 3.1.0
info:
  title: api.v1
paths:
  /api.v1.Service/CreateJobApplication:
    post:
      tags:
        - api.v1.Service
      summary: CreateJobApplication
      operationId: api.v1.Service.CreateJobApplication
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.v1.CreateJobApplicationRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.v1.CreateJobApplicationResponse'
  /api.v1.Service/ListJobApplications:
    post:
      tags:
        - api.v1.Service
      summary: ListJobApplications
      operationId: api.v1.Service.ListJobApplications
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.v1.ListJobApplicationsRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.v1.ListJobApplicationsResponse'
  /api.v1.Service/UpdateJobApplication:
    post:
      tags:
        - api.v1.Service
      summary: UpdateJobApplication
      operationId: api.v1.Service.UpdateJobApplication
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.v1.UpdateJobApplicationRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.v1.UpdateJobApplicationResponse'
  /api.v1.Service/DeleteJobApplication:
    post:
      tags:
        - api.v1.Service
      summary: DeleteJobApplication
      operationId: api.v1.Service.DeleteJobApplication
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.v1.DeleteJobApplicationRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.v1.DeleteJobApplicationResponse'
components:
  schemas:
    api.v1.JobApplicationStatus:
      type: string
      title: JobApplicationStatus
      enum:
        - JOB_APPLICATION_STATUS_UNSPECIFIED
        - JOB_APPLICATION_STATUS_APPLIED
        - JOB_APPLICATION_STATUS_SCREENING
        - JOB_APPLICATION_STATUS_INTERVIEW
        - JOB_APPLICATION_STATUS_OFFER
        - JOB_APPLICATION_STATUS_REJECTED
        - JOB_APPLICATION_STATUS_WITHDRAWN
        - JOB_APPLICATION_STATUS_ACCEPTED
    api.v1.CreateJobApplicationRequest:
      type: object
      properties:
        company:
          type: string
          title: company
        title:
          type: string
          title: title
        description:
          title: description
          $ref: '#/components/schemas/google.protobuf.StringValue'
        notes:
          title: notes
          $ref: '#/components/schemas/google.protobuf.StringValue'
        cv:
          title: cv
          $ref: '#/components/schemas/google.protobuf.StringValue'
        coverLetter:
          title: cover_letter
          $ref: '#/components/schemas/google.protobuf.StringValue'
        appliedOn:
          title: applied_on
          $ref: '#/components/schemas/google.protobuf.Timestamp'
        status:
          title: status
          $ref: '#/components/schemas/api.v1.JobApplicationStatus'
      title: CreateJobApplicationRequest
      additionalProperties: false
    api.v1.CreateJobApplicationResponse:
      type: object
      properties:
        jobApplication:
          title: job_application
          $ref: '#/components/schemas/api.v1.JobApplication'
      title: CreateJobApplicationResponse
      additionalProperties: false
    api.v1.DeleteJobApplicationRequest:
      type: object
      properties:
        id:
          type: string
          title: id
      title: DeleteJobApplicationRequest
      additionalProperties: false
    api.v1.DeleteJobApplicationResponse:
      type: object
      title: DeleteJobApplicationResponse
      additionalProperties: false
    api.v1.JobApplication:
      type: object
      properties:
        id:
          type: string
          title: id
        company:
          type: string
          title: company
        title:
          type: string
          title: title
        status:
          title: status
          $ref: '#/components/schemas/api.v1.JobApplicationStatus'
        description:
          title: description
          $ref: '#/components/schemas/google.protobuf.StringValue'
        notes:
          title: notes
          $ref: '#/components/schemas/google.protobuf.StringValue'
        cv:
          title: cv
          $ref: '#/components/schemas/google.protobuf.StringValue'
        coverLetter:
          title: cover_letter
          $ref: '#/components/schemas/google.protobuf.StringValue'
        appliedOn:
          title: applied_on
          $ref: '#/components/schemas/google.protobuf.Timestamp'
        createdAt:
          title: created_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
        updatedAt:
          title: updated_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
      title: JobApplication
      additionalProperties: false
    api.v1.ListJobApplicationsRequest:
      type: object
      title: ListJobApplicationsRequest
      additionalProperties: false
    api.v1.ListJobApplicationsResponse:
      type: object
      properties:
        jobApplications:
          type: array
          items:
            $ref: '#/components/schemas/api.v1.JobApplication'
          title: job_applications
      title: ListJobApplicationsResponse
      additionalProperties: false
    api.v1.UpdateJobApplicationRequest:
      type: object
      properties:
        id:
          type: string
          title: id
        company:
          type: string
          title: company
        title:
          type: string
          title: title
        description:
          title: description
          $ref: '#/components/schemas/google.protobuf.StringValue'
        notes:
          title: notes
          $ref: '#/components/schemas/google.protobuf.StringValue'
        cv:
          title: cv
          $ref: '#/components/schemas/google.protobuf.StringValue'
        coverLetter:
          title: cover_letter
          $ref: '#/components/schemas/google.protobuf.StringValue'
        status:
          title: status
          $ref: '#/components/schemas/api.v1.JobApplicationStatus'
        appliedOn:
          title: applied_on
          $ref: '#/components/schemas/google.protobuf.Timestamp'
      title: UpdateJobApplicationRequest
      additionalProperties: false
    api.v1.UpdateJobApplicationResponse:
      type: object
      properties:
        jobApplication:
          title: job_application
          $ref: '#/components/schemas/api.v1.JobApplication'
      title: UpdateJobApplicationResponse
      additionalProperties: false
    google.protobuf.StringValue:
      type: string
      description: |-
        Wrapper message for `string`.

         The JSON representation for `StringValue` is JSON string.

         Not recommended for use in new APIs, but still useful for legacy APIs and
         has no plan to be removed.
    google.protobuf.Timestamp:
      type: string
      examples:
        - 1s
        - 1.000340012s
      format: date-time
      description: |-
        A Timestamp represents a point in time independent of any time zone or local
         calendar, encoded as a count of seconds and fractions of seconds at
         nanosecond resolution. The count is relative to an epoch at UTC midnight on
         January 1, 1970, in the proleptic Gregorian calendar which extends the
         Gregorian calendar backwards to year one.

         All minutes are 60 seconds long. Leap seconds are "smeared" so that no leap
         second table is needed for interpretation, using a [24-hour linear
         smear](https://developers.google.com/time/smear).

         The range is from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59.999999999Z. By
         restricting to that range, we ensure that we can convert to and from [RFC
         3339](https://www.ietf.org/rfc/rfc3339.txt) date strings.

         # Examples

         Example 1: Compute Timestamp from POSIX `time()`.

             Timestamp timestamp;
             timestamp.set_seconds(time(NULL));
             timestamp.set_nanos(0);

         Example 2: Compute Timestamp from POSIX `gettimeofday()`.

             struct timeval tv;
             gettimeofday(&tv, NULL);

             Timestamp timestamp;
             timestamp.set_seconds(tv.tv_sec);
             timestamp.set_nanos(tv.tv_usec * 1000);

         Example 3: Compute Timestamp from Win32 `GetSystemTimeAsFileTime()`.

             FILETIME ft;
             GetSystemTimeAsFileTime(&ft);
             UINT64 ticks = (((UINT64)ft.dwHighDateTime) << 32) | ft.dwLowDateTime;

             // A Windows tick is 100 nanoseconds. Windows epoch 1601-01-01T00:00:00Z
             // is 11644473600 seconds before Unix epoch 1970-01-01T00:00:00Z.
             Timestamp timestamp;
             timestamp.set_seconds((INT64) ((ticks / 10000000) - 11644473600LL));
             timestamp.set_nanos((INT32) ((ticks % 10000000) * 100));

         Example 4: Compute Timestamp from Java `System.currentTimeMillis()`.

             long millis = System.currentTimeMillis();

             Timestamp timestamp = Timestamp.newBuilder().setSeconds(millis / 1000)
                 .setNanos((int) ((millis % 1000) * 1000000)).build();

         Example 5: Compute Timestamp from Java `Instant.now()`.

             Instant now = Instant.now();

             Timestamp timestamp =
                 Timestamp.newBuilder().setSeconds(now.getEpochSecond())
                     .setNanos(now.getNano()).build();

         Example 6: Compute Timestamp from current time in Python.

             timestamp = Timestamp()
             timestamp.GetCurrentTime()

         # JSON Mapping

         In JSON format, the Timestamp type is encoded as a string in the
         [RFC 3339](https://www.ietf.org/rfc/rfc3339.txt) format. That is, the
         format is "{year}-{month}-{day}T{hour}:{min}:{sec}[.{frac_sec}]Z"
         where {year} is always expressed using four digits while {month}, {day},
         {hour}, {min}, and {sec} are zero-padded to two digits each. The fractional
         seconds, which can go up to 9 digits (i.e. up to 1 nanosecond resolution),
         are optional. The "Z" suffix indicates the timezone ("UTC"); the timezone
         is required. A proto3 JSON serializer should always use UTC (as indicated by
         "Z") when printing the Timestamp type and a proto3 JSON parser should be
         able to accept both UTC and other timezones (as indicated by an offset).

         For example, "2017-01-15T01:30:15.01Z" encodes 15.01 seconds past
         01:30 UTC on January 15, 2017.

         In JavaScript, one can convert a Date object to this format using the
         standard
         [toISOString()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toISOString)
         method. In Python, a standard `datetime.datetime` object can be converted
         to this format using
         [`strftime`](https://docs.python.org/2/library/time.html#time.strftime) with
         the time format spec '%Y-%m-%dT%H:%M:%S.%fZ'. Likewise, in Java, one can use
         the Joda Time's [`ISODateTimeFormat.dateTime()`](
         http://joda-time.sourceforge.net/apidocs/org/joda/time/format/ISODateTimeFormat.html#dateTime()
         ) to obtain a formatter capable of generating timestamps in this format.
    connect-protocol-version:
      type: number
      title: Connect-Protocol-Version
      enum:
        - 1
      description: Define the version of the Connect protocol
      const: 1
    connect-timeout-header:
      type: number
      title: Connect-Timeout-Ms
      description: Define the timeout, in ms
    connect.error:
      type: object
      properties:
        code:
          type: string
          examples:
            - not_found
          enum:
            - canceled
            - unknown
            - invalid_argument
            - deadline_exceeded
            - not_found
            - already_exists
            - permission_denied
            - resource_exhausted
            - failed_precondition
            - aborted
            - out_of_range
            - unimplemented
            - internal
            - unavailable
            - data_loss
            - unauthenticated
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        detail:
          $ref: '#/components/schemas/google.protobuf.Any'
      title: Connect Error
      additionalProperties: true
      description: 'Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation'
    google.protobuf.Any:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
          format: binary
        debug:
          type: object
          additionalProperties: true
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
security: []
tags:
  - name: api.v1.Service